+++ members/nullnet-server/src/db/migrations/2026-09-16-000001_session_counts/down.sql
@@ -1,0 +1,5 @@
+DROP TRIGGER session_policy_update;
+DROP TRIGGER session_policy_delete;
+DROP TRIGGER session_policy_insert;
+DROP TABLE session_policy_counts;
+DROP INDEX sessions_active_counts_idx;